# contracts/api-contract.yaml
...
openapi: 3.0.3
info:
  title: kart-admin-service API
  version: 1.0.0
  description: >
    Back-office operations and fine-grained, category-scoped RBAC for the
    platform's four fixed operational categories (catalog management, coupon
    issuance, user suspension, inventory replenishment) plus a fifth
    meta-category (permission-management) that governs the other four
    (BRD §2.1 item 20, §24.1; ADR-0010). Admin Service is an orchestration/
    control-plane layer only — it never stores a second copy of Product,
    Category, Coupon, user-identity, or Inventory data (Domain Invariant #3);
    every mutation here is proxied synchronously to the owning service's own
    write API after Admin's own fine-grained grant check passes.
servers:
  - url: /v1
paths:
...
  /admin/attributes:
    post:
      operationId: createAttribute
      summary: Create an attribute via Category Service's own Attribute write API.
      security:
        - bearerAuth: [admin]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttributeWriteRequest' }
      responses:
        '201':
          description: Attribute created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminActionResult' }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Caller lacks the coarse `Admin` role claim, or lacks a live `catalog-management` grant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: categoryId not found or not active.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503':
          description: Category Service unavailable (circuit breaker open) — no write happened.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
  /admin/attributes/{attributeId}:
    put:
      operationId: updateAttribute
      summary: Rename an attribute and replace its values list via Category Service's own write API.
      security:
        - bearerAuth: [admin]
      parameters:
        - name: attributeId
          in: path
          required: true
          schema: { type: string }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttributeUpdateRequest' }
      responses:
        '200':
          description: Attribute updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminActionResult' }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Caller lacks the coarse `Admin` role claim, or lacks a live `catalog-management` grant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: attributeId not found or already deprecated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503':
          description: Category Service unavailable (circuit breaker open) — no write happened.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

    delete:
      operationId: deprecateAttribute
      summary: Deprecate an attribute via Category Service's own write API.
      security:
        - bearerAuth: [admin]
      parameters:
        - name: attributeId
          in: path
          required: true
          schema: { type: string }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Attribute deprecated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminActionResult' }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Caller lacks the coarse `Admin` role claim, or lacks a live `catalog-management` grant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: attributeId not found or already deprecated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503':
          description: Category Service unavailable (circuit breaker open) — no write happened.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ---------------------------------------------------------------------
  # Coupon issuance (category `coupon-issuance`). Proxies Offer Service's
  # POST /coupons and POST /coupons/{couponCode}/deactivate
  # (kart-offer-service/requirement-spec.md §5, "internal, admin-only" —
  # ADR-0010). Field names mirror docs/ddd/ubiquitous-language.md's
  # Offer-owned Coupon/CouponCode/RedemptionLimit definitions.
  # ---------------------------------------------------------------------
  /admin/coupons:
    post:
      operationId: createCoupon
      summary: Create a Coupon via Offer Service's admin-only write API.
      description: >
        Category `coupon-issuance`. Calls Offer Service's
        `POST /coupons` synchronously before committing Admin's own audit
        row (action `coupon.create`).
      security:
        - bearerAuth: [admin]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CouponWriteRequest' }
      responses:
        '201':
          description: Coupon created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminActionResult' }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Caller lacks the coarse `Admin` role claim, or lacks a live `coupon-issuance` grant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: couponCode already exists (Offer Service's `coupons.coupon_code` PRIMARY KEY).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503':
          description: Offer Service unavailable (circuit breaker open) — no write happened.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
  /admin/coupons/{couponCode}/deactivate:
    post:
      operationId: deactivateCoupon
      summary: Deactivate a Coupon via Offer Service's admin-only write API.
      description: >
        Deactivation only prevents future validate/redeem calls from seeing
        the code as active — it never retroactively invalidates a
        redemption/quote that already completed
        (kart-offer-service/requirement-spec.md, Domain Invariant on
        deactivation).
      security:
        - bearerAuth: [admin]
      parameters:
        - name: couponCode
          in: path
          required: true
          schema: { type: string }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Coupon deactivated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminActionResult' }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Caller lacks the coarse `Admin` role claim, or lacks a live `coupon-issuance` grant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: couponCode not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503':
          description: Offer Service unavailable (circuit breaker open) — no write happened.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ---------------------------------------------------------------------
  # User suspension (category `user-suspension`). Proxies Identity
  # Service's own internal, Admin-scoped routes — exact paths already
  # fixed by kart-identity-service/api-contract.yaml:
  # POST /internal/users/{userId}/lock and .../unlock.
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Identity-issued RS256 JWT carrying the `Admin` role claim
        (kart-identity-service/api-contract.yaml's TokenPair.roles), coarse-
        validated at the API Gateway before the request reaches Admin
        Service (BRD §24.1's two-tier enforcement, requirement-spec.md §2).
        Admin Service performs no separate coarse revocation-list check of
        its own (edge-cases.md, "Stale Admin Permission Outliving an
        Identity-Side Revocation") — it trusts the Gateway's per-request
        revalidation for this layer and adds only its own fine-grained,
        category-scoped grant check (admin_permission_grants) on top.
  schemas:
    GrantCategory:
      type: string
      description: >
        The six fixed categories requirement-spec.md §6 Decision item 1
        and database-design.md's CHECK constraint define — five business
        categories (`order-management` added for the "Order Management
        (Admin)" flow) plus the `permission-management` meta-category.
      enum:
        - catalog-management
        - coupon-issuance
        - user-suspension
        - inventory-replenishment
        - permission-management
        - order-management
...
    AdminActionResult:
      type: object
      description: >
        Mirrors database-design.md's admin_actions row and the
        AdminActionPerformed event payload it backs (event-contract.md).
      required: [actionId, adminId, category, action, entityId, performedAt]
      properties:
        actionId: { type: string, format: uuid }
        adminId: { type: string, description: "Identity-issued principal id of the acting admin. AdminActionPerformed payload field adminId." }
        category: { $ref: '#/components/schemas/GrantCategory' }
        action:
          type: string
          description: >
            Specific operation within the category, e.g. product.create,
            product.deactivate, category.reorder, coupon.create,
            coupon.deactivate, user.lock, user.unlock, inventory.replenish,
            grant.issue, grant.revoke (database-design.md). AdminActionPerformed
            payload field `action`.
        entityId:
          type: string
          description: >
            Id of the mutated entity in the owning service (or this
            service's own grant_id for permission-management actions).
            AdminActionPerformed payload field `entityId`.
        context:
          type: object
          nullable: true
          additionalProperties: true
          description: Optional richer audit detail beyond the three published event fields (database-design.md).
        performedAt: { type: string, format: date-time }
        publishedAt:
          type: string
          format: date-time
          nullable: true
          description: Set only once the Outbox poller has published AdminActionPerformed for this row (database-design.md).
...
    AttributeWriteRequest:
      type: object
      description: Mirrors Category Service's own POST /v1/attributes. categoryId null creates a global attribute.
      required: [name, dataType]
      properties:
        name: { type: string }
        categoryId: { type: string, nullable: true }
        dataType: { type: string, enum: [text, number, boolean, select] }
        values:
          type: array
          items: { $ref: '#/components/schemas/AttributeValueWriteRequest' }
    AttributeUpdateRequest:
      type: object
      description: Mirrors Category Service's own PATCH /v1/attributes/{id}. categoryId/dataType are immutable after creation.
      required: [name]
      properties:
        name: { type: string }
        values:
          type: array
          items: { $ref: '#/components/schemas/AttributeValueWriteRequest' }
    AttributeValueWriteRequest:
      type: object
      required: [value, displayOrder]
      properties:
        value: { type: string }
        displayOrder: { type: integer }
    CouponWriteRequest:
      type: object
      description: >
        Best-effort mirror of Offer Service's own POST /coupons request
        shape, using docs/ddd/ubiquitous-language.md's Offer-owned
        Coupon/CouponCode/RedemptionLimit field names.
      required: [couponCode, discountValue]
      properties:
        couponCode: { type: string }
        discountValue: { $ref: '#/components/schemas/Money' }
        perUserCap: { type: integer, nullable: true }
        globalCap: { type: integer, nullable: true }
        validFrom: { type: string, format: date-time, nullable: true }
        validUntil: { type: string, format: date-time, nullable: true }
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: number }
        currency: { type: string, example: USD }
    Problem:
      type: object
      description: RFC 7807-style problem details (matches kart-identity-service/api-contract.yaml's Problem convention).
      required: [code, message]
      properties:
        code:
          type: string
          example: permission_denied
        message:
          type: string
        details:
          type: object
          additionalProperties: true